+++ src/query_module.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#ifdef _WIN32
+
+#include "module_registry.h"
+
+HWND create_query_module(const ModuleContext& ctx);
+
+#endif