+++ src/settings_module.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#ifdef _WIN32
+
+#include "module_registry.h"
+
+HWND create_settings_module(const ModuleContext& ctx);
+
+#endif